refactor: preço deixou de ser double e passou a ser BigDecimal

+++ src/main/java/br/univ/pharmasys/model/Pedido.java
@@ -8,8 +8,8 @@ public class Pedido {
 	private int quantidade;
 	private Status status;
 	private LocalDate data;
-	private double preco_unitario;
-	private double valor_total;
+	private BigDecimal preco_unitario;
+	private BigDecimal valor_total;
 
 
 public  long getIdPedido() {
